% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_standard_error}
\alias{compute_standard_error}
\title{Compute Standard Error (SE) and Critical Difference (CD/LSD)}
\usage{
compute_standard_error(
  x = NULL,
  mean_square_error = NULL,
  total_replications = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A numeric vector of phenotypic observations (optional).}

\item{mean_square_error}{A numeric scalar representing MSE.}

\item{total_replications}{An integer scalar representing total replications.}

\item{na.rm}{Logical, default TRUE.}
}
\value{
A numeric scalar representing the calculated Standard Error.
}
\description{
Evaluated standard error (SE) of the mean and documented critical difference (CD) 
for phenotypic traits. Supports direct vector calculation for traits like Plant 
Height (PH) or ANOVA-based calculations using Mean Square Error (MSE).
}
\examples{
data(gv_data, package = "AgriDataTools")

# Example 1: Analyzed standard error for Plant Height (PH)
se_ph <- compute_standard_error(x = gv_data$PH)
print(paste("Standard Error for Plant Height:", round(se_ph, 4)))

}
\author{
Faheem Khan
}
