% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgriDataTools-package.R
\docType{package}
\name{AgriDataTools-package}
\alias{AgriDataTools}
\alias{AgriDataTools-package}
\title{AgriDataTools: Automated Statistical Analysis and Tools for Agricultural Research}
\description{
A comprehensive, high-performance suite of statistical tools tailored for agricultural and plant breeding research. Provides automated pipelines for analysis of variance (ANOVA) under Randomized Complete Block Designs (RCBD) and Completely Randomized Designs (CRD), variability parameters estimation, post-hoc multiple comparisons, multivariate analysis including principal component analysis (PCA), path analysis, and genotypic clustering.
}
\author{
\strong{Maintainer}: Faheem Khan \email{2022ag94@uaf.edu.pk} (\href{https://orcid.org/0009-0002-8963-7157}{ORCID})

Authors:
\itemize{
  \item Faheem Khan \email{2022ag94@uaf.edu.pk} (\href{https://orcid.org/0009-0002-8963-7157}{ORCID})
}

}
\keyword{internal}
